fix: harden money paths and close cross-platform parity gaps
Money-path hardening:
- getLatestNonce throws on an undecodable adminNonce response instead of
  silently signing withdrawals with nonce 0
- Portal rejects Solana chain ids on every EVM send/sign path via the new
  public RainChain.isSolana, instead of encoding eip155:900 transactions
- hexToBytes validates hex and enforces expected byte counts; withdrawal
  building requires 65-byte signatures and 32-byte salts
- toBaseUnits rejects negative amounts (RAIN_406) before uint256 encoding
- Solana withdrawals share the EVM withdrawal error contract: simulation
  reverts surface as RAIN_405 and raw parsing exceptions are wrapped

+++ rain-core-android/src/main/java/com/rain/sdk/internal/network/chainreader/ChainReader.kt
@@ -3,6 +3,7 @@ package com.rain.sdk.internal.network.chainreader
 import com.rain.sdk.models.Balance
 import com.rain.sdk.models.Token
 import com.rain.sdk.models.TokenInfo
+import java.math.BigDecimal
 
 /**
  * Provider-agnostic, read-only on-chain query surface.
@@ -24,7 +25,7 @@ internal interface ChainReader {
      * Native balance (e.g. ETH on Ethereum, AVAX on Avalanche). Result is in
      * human-readable form (e.g. `1.5` for 1.5 ETH).
      */
-    suspend fun getNativeBalance(chainId: Int, walletAddress: String): Double
+    suspend fun getNativeBalance(chainId: Int, walletAddress: String): BigDecimal
 
     /**
      * Single ERC-20 balance via `balanceOf(address)`. [decimals] defaults to
@@ -35,7 +36,7 @@ internal interface ChainReader {
         tokenAddress: String,
         walletAddress: String,
         decimals: Int?
-    ): Double
+    ): BigDecimal
 
     /**
      * Batched balances for many tokens on one chain, in a single round-trip when possible.
